[ci skip] Delete SwiftTunnelHooks, use observable as hooks
