+++ app-apple/Sources/AppLibrary/Previews/AppContext+Previews.swift
@@ -52,7 +52,7 @@ extension AppContext {
         ) { @Sendable _ in
             SharedTunnelEnvironment(profileId: nil)
         }
-        let tunnelHooks = SwiftTunnelHooks(tunnel: tunnel)
+        let tunnelObservable = TunnelObservable(tunnel: tunnel)
         let preferencesManager = PreferencesManager()
 
         let dummyReceiver = DummyWebReceiver(url: URL(string: "http://127.0.0.1:9000")!)
@@ -75,7 +75,7 @@ extension AppContext {
             preferencesManager: preferencesManager,
             profileManager: profileManager,
             registry: registry,
-            tunnelHooks: tunnelHooks,
+            tunnelHooks: tunnelObservable,
             versionChecker: versionChecker,
             webReceiverManager: webReceiverManager,
             bindings: nil
@@ -84,7 +84,7 @@ extension AppContext {
             abi: abi,
             appConfiguration: appConfiguration,
             kvStore: kvStore,
-            tunnelObservable: TunnelObservable(tunnel: tunnel)
+            tunnelObservable: tunnelObservable
         )
     }()
 }
